# src/main/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HeldContentAnswer": {
        "description": "What the gateway will say about one pair a client holds",
        "properties": {
          "approvedOverReversedRevocation": {
            "description": "True when this approval stands over an administrative revocation somebody reversed, so content served over a reversed withdrawal is never indistinguishable from content nobody ever withdrew",
            "type": "boolean"
          },
          "marketplace": {
            "description": "Marketplace name as asked",
            "type": "string"
          },
          "revokedAt": {
            "description": "When the content was withdrawn; null unless the state is 'revoked'",
            "format": "date-time",
            "type": "string"
          },
          "sha": {
            "description": "Commit as asked",
            "type": "string"
          },
          "state": {
            "description": "approved, revoked, or unknown. 'unknown' is not a pass: the gateway makes no positive statement, and a holder should treat the content as withdrawn.",
            "enum": [
              "approved",
              "revoked",
              "unknown"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "HeldContentQuery": {
        "description": "One marketplace-and-commit pair a client holds",
        "properties": {
          "marketplace": {
            "description": "Marketplace name, as it appears in the fetch URL",
            "type": "string"
          },
          "sha": {
            "description": "The commit the client holds",
            "type": "string"
          }
        },
        "type": "object"
      },
      "HeldContentReport": {
        "description": "One state per requested pair, in the order asked",
        "properties": {
          "results": {
            "description": "The answers",
            "items": {
              "$ref": "#/components/schemas/HeldContentAnswer"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "HeldContentRequest": {
        "description": "The content a client holds and wants a statement about",
        "properties": {
          "holdings": {
            "description": "Marketplace-and-commit pairs the client holds, at most 256",
            "items": {
              "$ref": "#/components/schemas/HeldContentQuery"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "PlaygroundRequest": {
        "description": "Playground evaluation request",
        "properties": {
          "expression": {
            "description": "CEL expression to test; it is compiled and evaluated but never stored",
            "example": "skills.exists(s, s.tools.exists(t, t.startsWith(\"Bash\")))",
            "type": "string"
          },
          "snapshotId": {
            "description": "The real snapshot to evaluate against",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PlaygroundResult": {
        "description": "Playground answer: what the expression said, or why it could not say anything",
        "properties": {
          "error": {
            "description": "The compile or evaluation error; absent when the expression answered",
            "type": "string"
          },
          "matched": {
            "description": "Whether the expression matched; absent when it errored",
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    }
  },
...
  "info": {
    "contact": {
      "name": "jimisola",
      "url": "https://github.com/skillsgateway/skillsgateway"
    },
...
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Skills Gateway",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/policy/playground": {
      "post": {
        "description": "Compiles and evaluates any CEL expression over a real snapshot's facts — held, approved or revoked — and answers matched or the error. Read-only by contract: nothing is stored, nothing lands on the ledger, no state changes; the answer never carries snapshot content. This is how a rule is tested before it is enforced. Requires permission to approve the named snapshot.",
        "operationId": "playground",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaygroundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundResult"
                }
              }
            },
            "description": "The expression's answer, or its error"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundResult"
                }
              }
            },
            "description": "No such snapshot"
          }
        },
        "summary": "Test an expression against a real snapshot",
        "tags": [
          "Policy"
        ]
      }
    },
...
    "/status/v1/snapshots": {
      "post": {
        "description": "Answers one state per marketplace-and-commit pair, in the order asked. A pair the gateway makes no positive statement about is answered 'unknown', which is never a pass: a client that holds such content should treat it as withdrawn. Authenticates with the same credentials the git facade accepts, and answers only about marketplaces that credential could fetch.",
        "operationId": "status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeldContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "One state per requested pair, in the order asked"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "Malformed request, or more pairs than one request may carry"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HeldContentReport"
                }
              }
            },
            "description": "Unauthenticated"
          }
        },
        "summary": "Ask whether content a client already holds is still approved",
        "tags": [
          "Held content"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}